Refactored params package structure
