+++ modules/nextflow/src/main/groovy/nextflow/processor/TaskProcessor.groovy
@@ -63,24 +63,24 @@ import nextflow.file.FileHelper
 import nextflow.file.FileHolder
 import nextflow.file.FilePatternSplitter
 import nextflow.script.BaseScript
-import nextflow.script.BasicMode
-import nextflow.script.EachInParam
-import nextflow.script.EnvInParam
-import nextflow.script.FileInParam
-import nextflow.script.FileOutParam
-import nextflow.script.InParam
-import nextflow.script.MissingParam
-import nextflow.script.OptionalParam
-import nextflow.script.OutParam
+import nextflow.script.params.BasicMode
+import nextflow.script.params.EachInParam
+import nextflow.script.params.EnvInParam
+import nextflow.script.params.FileInParam
+import nextflow.script.params.FileOutParam
+import nextflow.script.params.InParam
+import nextflow.script.params.MissingParam
+import nextflow.script.params.OptionalParam
+import nextflow.script.params.OutParam
 import nextflow.script.ProcessConfig
 import nextflow.script.ScriptType
-import nextflow.script.SetInParam
-import nextflow.script.SetOutParam
-import nextflow.script.StdInParam
-import nextflow.script.StdOutParam
+import nextflow.script.params.SetInParam
+import nextflow.script.params.SetOutParam
+import nextflow.script.params.StdInParam
+import nextflow.script.params.StdOutParam
 import nextflow.script.TaskBody
-import nextflow.script.ValueInParam
-import nextflow.script.ValueOutParam
+import nextflow.script.params.ValueInParam
+import nextflow.script.params.ValueOutParam
 import nextflow.util.ArrayBag
 import nextflow.util.BlankSeparatedList
 import nextflow.util.CacheHelper
