+++ modules/nextflow/src/main/groovy/nextflow/processor/TaskRun.groovy
@@ -34,15 +34,15 @@ import nextflow.exception.ProcessTemplateException
 import nextflow.exception.ProcessUnrecoverableException
 import nextflow.file.FileHelper
 import nextflow.file.FileHolder
-import nextflow.script.EnvInParam
-import nextflow.script.FileInParam
-import nextflow.script.FileOutParam
-import nextflow.script.InParam
-import nextflow.script.OutParam
+import nextflow.script.params.EnvInParam
+import nextflow.script.params.FileInParam
+import nextflow.script.params.FileOutParam
+import nextflow.script.params.InParam
+import nextflow.script.params.OutParam
 import nextflow.script.ScriptType
-import nextflow.script.StdInParam
+import nextflow.script.params.StdInParam
 import nextflow.script.TaskBody
-import nextflow.script.ValueOutParam
+import nextflow.script.params.ValueOutParam
 /**
  * Models a task instance
  *
