+++ modules/nextflow/src/test/groovy/nextflow/processor/TaskProcessorTest.groovy
@@ -31,7 +31,7 @@ import nextflow.exception.ProcessUnrecoverableException
 import nextflow.executor.NopeExecutor
 import nextflow.file.FileHolder
 import nextflow.script.BaseScript
-import nextflow.script.FileOutParam
+import nextflow.script.params.FileOutParam
 import nextflow.script.ProcessConfig
 import nextflow.script.ScriptType
 import nextflow.script.TaskBody
